:recycle: Move types used by item-sliding to appropiate file
